+++ config/globals.js
@@ -42,10 +42,6 @@ const globals = {
 	MAXIMUM_ALLOWABLE_TOTAL_ATTACHMENTS_SIZE_BYTE: 5 * 1024 * 1024, // 5MB
 	MONGOOSE_CONNECTION_POOL_SIZE: parseInt(process.env.MONGOOSE_CONNECTION_POOL_SIZE || '10', 10),
 
-	SC_DOMAIN: process.env.SC_DOMAIN || 'localhost',
-	SC_THEME: process.env.SC_THEME || 'default',
-	SC_TITLE: process.env.SC_TITLE || 'HPI Schul-Cloud',
-	SC_SHORT_TITLE: process.env.SC_SHORT_TITLE || 'HPI Schul-Cloud',
 	SMTP_SENDER: process.env.SMTP_SENDER || 'noreply@schul-cloud.org',
 
 	KEEP_ALIVE: process.env.KEEP_ALIVE || false,