+++ src/services/teams/hooks/index.js
@@ -1,3 +1,5 @@
+const { Configuration } = require('@hpi-schul-cloud/commons');
+
 const { authenticate } = require('@feathersjs/authentication');
 const reqlib = require('app-root-path').require;
 
@@ -5,7 +7,6 @@ const { Forbidden, NotFound, BadRequest, Conflict, NotImplemented, MethodNotAllo
 	'src/errors'
 );
 const { equal: equalIds } = require('../../../helper/compare').ObjectId;
-const { SC_SHORT_TITLE } = require('../../../../config/globals');
 
 const globalHooks = require('../../../hooks');
 const logger = require('../../../logger');
@@ -593,7 +594,7 @@ const sendInfo = (hook) => {
 	return getSessionUser(hook)
 		.then((user) => {
 			globalHooks.sendEmail(hook, {
-				subject: `${SC_SHORT_TITLE}: Team-Einladung`,
+				subject: `${Configuration.get('SC__SHORT_TITLE')}: Team-Einladung`,
 				emails: [email],
 				content: {
 					text: createEmailText(hook, user),