+++ test/services/sync/strategies/CSVSyncer/CSVSyncer.integration.test.js
@@ -4,6 +4,7 @@ const appPromise = require('../../../../../src/app');
 const roleModel = require('../../../../../src/services/role/model.js');
 const { userModel } = require('../../../../../src/services/user/model');
 const MailService = require('../../../../../src/services/helpers/service.js');
+const { Configuration } = require('@hpi-schul-cloud/commons');
 
 const testObjects = require('../../../helpers/testObjects')(appPromise);
 const { generateRequestParamsFromUser } = require('../../../helpers/services/login')(appPromise);
@@ -18,7 +19,6 @@ const {
 } = require('../../../helpers/services/classes')(appPromise);
 
 const CSVSyncer = require('../../../../../src/services/sync/strategies/CSVSyncer');
-const { SC_TITLE } = require('../../../../../config/globals');
 
 const { deleteUser, MockEmailService } = require('./helper');
 
@@ -482,7 +482,7 @@ describe('CSVSyncer Integration', () => {
 			expect(stats.classes.failed).to.equal(0);
 
 			expect(emails.length).to.equal(3);
-			expect(emails[0].subject).to.equal(`Einladung für die Nutzung der ${SC_TITLE}!`);
+			expect(emails[0].subject).to.equal(`Einladung für die Nutzung der ${Configuration.get('SC__TITLE')}!`);
 			expect(emails[0].content.text).to.include('Hallo Chuck Bartowski!');
 
 			await Promise.all(